extract substep logic from steps to substeps.go
